+++ lib/cached/md5_cached.cmake
@@ -1,4 +1,4 @@
 set(lexer_lxx_md5_cached "29c69cfc368018c0bf8fe3d63f00c677")
-set(parser_yxx_md5_cached "1f9d72d236e240070ee3fab3d6fe432e")
+set(parser_yxx_md5_cached "37a95fccfb5811313546809fe8d85a2a")
 set(regex_lexer_lxx_md5_cached "5b4c138ececa989083ac2ed39f8959c2")
 set(regex_parser_yxx_md5_cached "e44695248dcfad64a56982f126f8e4f5")